moved registry into generate
