+++ cmd/verify/verify.go
@@ -9,7 +9,7 @@ import (
 
 	cmd_generate "github.com/safe-waters/docker-lock/cmd/generate"
 	"github.com/safe-waters/docker-lock/pkg/generate"
-	"github.com/safe-waters/docker-lock/pkg/registry"
+	"github.com/safe-waters/docker-lock/pkg/generate/registry"
 	"github.com/safe-waters/docker-lock/pkg/verify"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
