+++ pkg/generate/registry/auth.go
