+++ pkg/generate/registry/helpers.go
