+++ pkg/generate/registry/manager.go
