+++ pkg/generate/registry/v2.go
