+++ pkg/generate/registry/wrapper.go
